Day_83/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.1.2
gunicorn==21.2.0
Werkzeug==2.3.7
python-dotenv==1.0.1
